+++ NEWS.md
@@ -8,6 +8,7 @@ New Features:
 
 Bug fixes:
 + Fix false positive `PhanParamSignatureMismatch` for variadic overriding a function using `func_get_args()` (#4340)
++ Don't emit PhanTypeNoPropertiesForeach for the Countable interface on its own. (#4342)
 
 Dec 31 2020, Phan 4.0.1
 -----------------------